Changed velocity delay type from float to int

+++ Loader/src/global_settings.h
@@ -196,7 +196,7 @@ namespace toad
 		inline bool kite = false; // don't decrease velocity when hit from behind
 
 		inline int chance = 100;
-		inline float delay = 0;
+		inline int delay = 0;
 
 		// in % the lower the less vel
 		inline float horizontal = 100;
